pyproject.toml
[project]
name = "socialpulse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.3",
    "black>=26.1.0",
    "fastapi>=0.135.0",
    "httpx>=0.28.1",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.47",
    "statsmodels>=0.14.6",
    "uvicorn>=0.41.0",
    "vadersentiment>=3.3.2",
]
